[Security Solution] move detection_engine file out of the detections folder
