+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/common/components/related_integrations/integrations_description/index.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import type { RelatedIntegrationArray } from '../../../../../../common/api/detection_engine/model/rule_schema';
-import type { ListItems } from '../../../../../detection_engine/rule_creation_ui/components/description_step/types';
+import type { ListItems } from '../../../../rule_creation_ui/components/description_step/types';
 import type { IntegrationDetails } from '../integration_details';
 import { useRelatedIntegrations } from '../use_related_integrations';
