+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/common/components/related_integrations/integrations_popover/index.tsx
