+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/common/components/related_integrations/use_integrations.tsx
@@ -8,7 +8,7 @@
 import { useQuery } from '@tanstack/react-query';
 
 import type { Integration } from '../../../../../common/api/detection_engine/fleet_integrations';
-import { fleetIntegrationsApi } from '../../../../detection_engine/fleet_integrations';
+import { fleetIntegrationsApi } from '../../../fleet_integrations';
 
 const ONE_MINUTE = 60000;
