+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/common/components/rule_execution_status/utils.ts
