+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/common/components/rule_switch/translations.ts
