+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_creation/components/related_integrations/related_integration_field.tsx
@@ -12,17 +12,17 @@ import semver from 'semver';
 import { css } from '@emotion/css';
 import type { EuiComboBoxOptionOption } from '@elastic/eui';
 import {
-  EuiTextTruncate,
   EuiButtonIcon,
   EuiComboBox,
   EuiFieldText,
   EuiFlexGroup,
   EuiFlexItem,
   EuiFormRow,
+  EuiTextTruncate,
 } from '@elastic/eui';
 import type { FieldHook } from '../../../../shared_imports';
 import type { Integration, RelatedIntegration } from '../../../../../common/api/detection_engine';
-import { useIntegrations } from '../../../../detections/components/rules/related_integrations/use_integrations';
+import { useIntegrations } from '../../../common/components/related_integrations/use_integrations';
 import { IntegrationStatusBadge } from './integration_status_badge';
 import * as i18n from './translations';
