+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/three_way_diff/final_readonly/fields/related_integrations/related_integrations.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import { EuiDescriptionList } from '@elastic/eui';
 import * as ruleDetailsI18n from '../../../../translations';
-import { RelatedIntegrationsDescription } from '../../../../../../../../detections/components/rules/related_integrations/integrations_description';
+import { RelatedIntegrationsDescription } from '../../../../../../../common/components/related_integrations/integrations_description';
 import type { RelatedIntegrationArray } from '../../../../../../../../../common/api/detection_engine';
 import { EmptyFieldValuePlaceholder } from '../../empty_field_value_placeholder';
