+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management_ui/components/background_image_callout/index.tsx
