+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management_ui/components/blog_post_prebuilt_rule_customization_callout/translations.ts
