+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management_ui/components/rules_table/rules_tables.tsx
@@ -16,7 +16,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import React, { useCallback, useMemo, useRef } from 'react';
 import { Loader } from '../../../../common/components/loader';
 import { useBoolState } from '../../../../common/hooks/use_bool_state';
-import { PrePackagedRulesPrompt } from '../../../../detections/components/rules/pre_packaged_rules/load_empty_prompt';
+import { PrePackagedRulesPrompt } from '../pre_packaged_rules/load_empty_prompt';
 import type { Rule } from '../../../rule_management/logic';
 import * as i18n from '../../../common/translations';
 import type { EuiBasicTableOnChange } from '../../../common/types';
