+++ x-pack/solutions/security/plugins/security_solution/public/flyout/entity_details/shared/hooks/use_managed_user.test.ts
@@ -26,9 +26,12 @@ const mockUseIntegrations = jest.fn().mockReturnValue({
   data: [],
 });
 
-jest.mock('../../../../detections/components/rules/related_integrations/use_integrations', () => ({
-  useIntegrations: () => mockUseIntegrations(),
-}));
+jest.mock(
+  '../../../../detection_engine/common/components/related_integrations/use_integrations',
+  () => ({
+    useIntegrations: () => mockUseIntegrations(),
+  })
+);
 
 jest.mock('../../../../common/hooks/use_space_id', () => ({
   useSpaceId: () => 'test-space-id',
