+++ x-pack/solutions/security/plugins/security_solution/public/siem_migrations/rules/components/rules_table_columns/integrations.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { EuiHorizontalRule, EuiLoadingSpinner, EuiText } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
 import type { RelatedIntegration } from '../../../../../common/api/detection_engine';
-import { IntegrationsPopover } from '../../../../detections/components/rules/related_integrations/integrations_popover';
+import { IntegrationsPopover } from '../../../../detection_engine/common/components/related_integrations/integrations_popover';
 import type { RuleMigrationRule } from '../../../../../common/siem_migrations/model/rule_migration.gen';
 import * as i18n from './translations';
 import type { TableColumn } from './constants';
